+++ packages/frontend/src/containers/Search.ts
@@ -1,0 +1,18 @@
+import { Container } from "unstated";
+
+interface ISearchContainerProps {
+  query: string;
+}
+
+class SearchContainer extends Container<ISearchContainerProps> {
+  state = {
+    query: ""
+  };
+
+  setQuery = (query: string) =>
+    this.setState({
+      query
+    });
+}
+
+export { SearchContainer };