+++ packages/frontend/src/downstream/Context/Input/Input.tsx
@@ -1,6 +1,14 @@
+import styled from "styled-components";
+import { Input as InputTemplate } from "semantic-ui-react";
+
 interface IInput {
   value: string;
   onChange: any;
 }
 
-export { IInput };
+const Input = styled(InputTemplate)`
+  width: 100%;
+  margin-right: 1em;
+`;
+
+export { IInput, Input };