+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## <small>0.0.1-0-monorepo (2018-12-21)</small>
 
+* refactor(frontend): Add better layout for context ([b4442fe](https://gitlab.com/pojntfx/learn-chinese-platform/commit/b4442fe))
 * refactor(frontend): Add refactored header to context ([9a22329](https://gitlab.com/pojntfx/learn-chinese-platform/commit/9a22329))
 * feat(frontend): Continue refactoring dictionary ([4e30bf7](https://gitlab.com/pojntfx/learn-chinese-platform/commit/4e30bf7))
 * feat(frontend): Start refactoring character ([217cd78](https://gitlab.com/pojntfx/learn-chinese-platform/commit/217cd78))